_pipelines/topics-v1/requirements.txt
# Topic Analysis Requirements

## Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

## Database & API
qdrant-client>=1.6.0
python-dotenv>=1.0.0

## Advanced Topic Modeling
umap-learn>=0.5.0
hdbscan>=0.8.33

## Optional (if using full BERTopic)
# bertopic>=0.16.0
# sentence-transformers>=2.2.0

## Installation
```bash
pip install -r requirements.txt
```

## Notes
- All embeddings are pre-computed in Qdrant (no need for sentence-transformers)
- UMAP provides better dimensionality reduction than PCA for topic discovery
- HDBSCAN with smart parameters handles varying community sizes
- Matplotlib backend issues may require: `export MPLBACKEND=Agg`

## Verified Working Versions
- Python 3.11 (via conda environment core311)
- macOS compatible (tested on M1 MacBook Air)
- Memory efficient for 5000+ document analysis
